requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
Pillow
opencv-python
numpy
pytesseract
reportlab
